Fix incorrect use of pubkey ptr in new_lease()
Before this, leases_ht would store the (temporary) pointer to pubkey
that's given to new_lease() directly. The memory it's pointing to is
overwritten as soon as a new connection comes in however and thus
breaking the hashtable.

Instead make our own copy of the pubkey and store the pointer to that.