radix-trie: fix issues related to shadow-/counting
Previously it was possible that pools would not be correctly shadowed
and/or the left or right counters were not updated properly. To fix
that, every node now has flags indicating what type of node it is, and
if it's shadowed. Furthermore, We determine if a poolnode needs to be
(un)shadowed by walking the trie now.

remove_node() now also only modifies the left right counters if the
target node was actually found.